refactor routing and middleware
