--- packages/open-next/src/adapters/plugins/routing/default.replacement.ts
@@ -1,154 +1,0 @@
-/* eslint-disable simple-import-sort/imports */
-import type {
-  PostProcessOptions,
-  ProcessInternalEvent,
-} from "../../types/plugin";
-import type { InternalResult } from "../../event-mapper";
-//#override imports
-
-import { debug } from "../../logger";
-import { IncomingMessage } from "../../http/request";
-import {
-  addNextConfigHeaders,
-  fixDataPage,
-  handleFallbackFalse,
-  handleRedirects,
-  handleRewrites,
-} from "../../routing/matcher";
-import {
-  addOpenNextHeader,
-  fixCacheHeaderForHtmlPages,
-  fixISRHeaders,
-  fixSWRCacheHeader,
-  revalidateIfRequired,
-} from "./util";
-import { convertRes } from "../../routing/util";
-import { handleMiddleware } from "../../routing/middleware";
-import {
-  BuildId,
-  ConfigHeaders,
-  PrerenderManifest,
-  RoutesManifest,
-} from "../../config";
-
-//#endOverride
-
-//#override processInternalEvent
-export const processInternalEvent: ProcessInternalEvent = async (
-  event,
-  createResponse,
-) => {
-  const nextHeaders = addNextConfigHeaders(event, ConfigHeaders) ?? {};
-
-  let internalEvent = fixDataPage(event, BuildId);
-
-  internalEvent = handleFallbackFalse(internalEvent, PrerenderManifest);
-
-  const redirect = handleRedirects(internalEvent, RoutesManifest.redirects);
-  if (redirect) {
-    return redirect;
-  }
-
-  const middleware = await handleMiddleware(internalEvent);
-  let middlewareResponseHeaders: Record<string, string | string[]> = {};
-  if ("statusCode" in middleware) {
-    return middleware;
-  } else {
-    middlewareResponseHeaders = middleware.responseHeaders || {};
-    internalEvent = middleware;
-  }
-
-  let isExternalRewrite = middleware.externalRewrite ?? false;
-  if (!isExternalRewrite) {
-    // First rewrite to be applied
-    const beforeRewrites = handleRewrites(
-      internalEvent,
-      RoutesManifest.rewrites.beforeFiles,
-    );
-    internalEvent = beforeRewrites.internalEvent;
-    isExternalRewrite = beforeRewrites.isExternalRewrite;
-  }
-  const isStaticRoute = RoutesManifest.routes.static.some((route) =>
-    new RegExp(route.regex).test(event.rawPath),
-  );
-
-  if (!isStaticRoute && !isExternalRewrite) {
-    // Second rewrite to be applied
-    const afterRewrites = handleRewrites(
-      internalEvent,
-      RoutesManifest.rewrites.afterFiles,
-    );
-    internalEvent = afterRewrites.internalEvent;
-    isExternalRewrite = afterRewrites.isExternalRewrite;
-  }
-
-  const isDynamicRoute = RoutesManifest.routes.dynamic.some((route) =>
-    new RegExp(route.regex).test(event.rawPath),
-  );
-  if (!isDynamicRoute && !isStaticRoute && !isExternalRewrite) {
-    // Fallback rewrite to be applied
-    const fallbackRewrites = handleRewrites(
-      internalEvent,
-      RoutesManifest.rewrites.fallback,
-    );
-    internalEvent = fallbackRewrites.internalEvent;
-    isExternalRewrite = fallbackRewrites.isExternalRewrite;
-  }
-
-  const reqProps = {
-    method: internalEvent.method,
-    url: internalEvent.url,
-    //WORKAROUND: We pass this header to the serverless function to mimic a prefetch request which will not trigger revalidation since we handle revalidation differently
-    // There is 3 way we can handle revalidation:
-    // 1. We could just let the revalidation go as normal, but due to race condtions the revalidation will be unreliable
-    // 2. We could alter the lastModified time of our cache to make next believe that the cache is fresh, but this could cause issues with stale data since the cdn will cache the stale data as if it was fresh
-    // 3. OUR CHOICE: We could pass a purpose prefetch header to the serverless function to make next believe that the request is a prefetch request and not trigger revalidation (This could potentially break in the future if next changes the behavior of prefetch requests)
-    headers: { ...internalEvent.headers, purpose: "prefetch" },
-    body: internalEvent.body,
-    remoteAddress: internalEvent.remoteAddress,
-  };
-  debug("IncomingMessage constructor props", reqProps);
-  const req = new IncomingMessage(reqProps);
-  const res = createResponse(reqProps.method, {
-    ...nextHeaders,
-    ...middlewareResponseHeaders,
-  });
-
-  return { internalEvent: internalEvent, req, res, isExternalRewrite };
-};
-//#endOverride
-
-//#override postProcessResponse
-export async function postProcessResponse({
-  internalEvent,
-  req,
-  res,
-  isExternalRewrite,
-}: PostProcessOptions): Promise<InternalResult> {
-  const { statusCode, headers, isBase64Encoded, body } = convertRes(res);
-
-  debug("ServerResponse data", { statusCode, headers, isBase64Encoded, body });
-
-  if (!isExternalRewrite) {
-    fixCacheHeaderForHtmlPages(internalEvent.rawPath, headers);
-    fixSWRCacheHeader(headers);
-    addOpenNextHeader(headers);
-    fixISRHeaders(headers);
-
-    await revalidateIfRequired(
-      internalEvent.headers.host,
-      internalEvent.rawPath,
-      headers,
-      req,
-    );
-  }
-
-  return {
-    type: internalEvent.type,
-    statusCode,
-    headers,
-    body,
-    isBase64Encoded,
-  };
-}
-//#endOverride
