+++ packages/open-next/src/adapters/plugins/serverHandler.replacement.ts
@@ -1,17 +1,17 @@
 /*eslint-disable simple-import-sort/imports */
 import type { Options, PluginHandler } from "../types/next-types.js";
 import type { IncomingMessage } from "../http/request.js";
-import type { ServerlessResponse } from "../http/response.js";
 //#override imports
 
 import { proxyRequest } from "./routing/util.js";
 import { requestHandler, setNextjsPrebundledReact } from "./util.js";
+import { OpenNextNodeResponse } from "../http/openNextResponse.js";
 //#endOverride
 
 //#override handler
 export const handler: PluginHandler = async (
   req: IncomingMessage,
-  res: ServerlessResponse,
+  res: OpenNextNodeResponse,
   options: Options,
 ) => {
   let { internalEvent } = options;
