+++ packages/open-next/src/adapters/plugins/serverHandler.ts
@@ -1,5 +1,5 @@
+import type { OpenNextNodeResponse } from "../http/openNextResponse.js";
 import type { IncomingMessage } from "../http/request.js";
-import { ServerlessResponse } from "../http/response.js";
 import type { Options, PluginHandler } from "../types/next-types.js";
 //#override imports
 import { requestHandler, setNextjsPrebundledReact } from "./util.js";
@@ -8,7 +8,7 @@ import { requestHandler, setNextjsPrebundledReact } from "./util.js";
 //#override handler
 export const handler: PluginHandler = async (
   req: IncomingMessage,
-  res: ServerlessResponse,
+  res: OpenNextNodeResponse,
   options: Options,
 ) => {
   setNextjsPrebundledReact(options.internalEvent.rawPath);
