+++ packages/open-next/src/adapters/types/next-types.ts
@@ -1,6 +1,7 @@
 // NOTE: add more next config typings as they become relevant
 
 import { InternalEvent } from "../event-mapper.js";
+import { OpenNextNodeResponse } from "../http/openNextResponse.js";
 import { IncomingMessage } from "../http/request.js";
 import { ServerlessResponse } from "../http/response.js";
 
@@ -152,7 +153,7 @@ export type Options = {
 export interface PluginHandler {
   (
     req: IncomingMessage,
-    res: ServerlessResponse,
+    res: OpenNextNodeResponse,
     options: Options,
   ): Promise<ServerlessResponse | undefined>;
 }
